Phase 1: extract in-document search into features/search.js
The search feature's state (matches, current index, highlight nodes) is used
only by search functions, so it moves into the module as private state — no
shared-state refactor needed. DOM elements are looked up by id inside the
module (decoupled from main.js's cached element consts). main.js imports the
public API (openSearch/closeSearch/runSearch/navigateSearch/
clearSearchHighlights); escapeRegex and the highlight/count helpers are private.

Pure refactor. main.js ~2,706 -> ~2,566 lines. Build + lint clean; 297 tests pass.
